+++ include/AST.hpp
@@ -87,6 +87,7 @@ public:
     std::string type;
     int pointerLevel;
     std::string name;
+    bool isVolatile = false;
     std::unique_ptr<Expression> initializer;
     VariableDeclaration(const std::string& t, const std::string& n, int p = 0) : type(t), pointerLevel(p), name(n) {}
     void accept(ASTVisitor& visitor) override;
@@ -176,6 +177,7 @@ struct Parameter {
     std::string type;
     int pointerLevel;
     std::string name;
+    bool isVolatile = false; // New: volatile qualifier
 };
 
 class FunctionDeclaration : public ASTNode {